fix(data): store the created timestamp as a string
Gorm doesn’t know how to properly marshal Timestamp, which is a
(database/sql/driver).Valuer, into a new cluster checkin row. so, we
have to store it as a string and manually marshal it.

this commit also adds gorm callbacks to enforce that the marshaled
value is a valid timestamp